# server/internal/apigen/apigen.gen.go
// Package apigen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package apigen

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
)

// Defines values for HealthStatus.
const (
	HealthStatusDown HealthStatus = "down"
	HealthStatusUp   HealthStatus = "up"
)

// Valid indicates whether the value is a known member of the HealthStatus enum.
func (e HealthStatus) Valid() bool {
	switch e {
	case HealthStatusDown:
		return true
	case HealthStatusUp:
		return true
	default:
		return false
	}
}

// Defines values for HealthCheckStatus.
const (
	HealthCheckStatusDown HealthCheckStatus = "down"
	HealthCheckStatusUp   HealthCheckStatus = "up"
)

// Valid indicates whether the value is a known member of the HealthCheckStatus enum.
func (e HealthCheckStatus) Valid() bool {
	switch e {
	case HealthCheckStatusDown:
		return true
	case HealthCheckStatusUp:
		return true
	default:
		return false
	}
}

// EnvelopeBase defines model for EnvelopeBase.
type EnvelopeBase struct {
	Error *ErrorBody `json:"error,omitempty"`
	Meta  *PageMeta  `json:"meta,omitempty"`

	// RequestId 请求 ID，与响应头 X-Request-ID 一致，便于排查
	RequestId string `json:"requestId"`
	Success   bool   `json:"success"`
}

// ErrorBody defines model for ErrorBody.
type ErrorBody struct {
	// Code 机器可读错误码
	//
	// Example: VALIDATION_FAILED
	Code string `json:"code"`

	// Details 附加信息，如字段级校验错误
	Details *map[string]interface{} `json:"details,omitempty"`

	// Message 面向用户的错误描述
	//
	// Example: 参数校验失败
	Message string `json:"message"`
}

// ErrorEnvelope defines model for ErrorEnvelope.
type ErrorEnvelope struct {
	Error ErrorBody `json:"error"`
	Meta  *PageMeta `json:"meta,omitempty"`

	// RequestId 请求 ID，与响应头 X-Request-ID 一致，便于排查
	RequestId string `json:"requestId"`
	Success   bool   `json:"success"`
}

// Health defines model for Health.
type Health struct {
	Checks []HealthCheck `json:"checks"`
	Status HealthStatus  `json:"status"`
}

// HealthStatus defines model for Health.Status.
type HealthStatus string

// HealthCheck defines model for HealthCheck.
type HealthCheck struct {
	Error *string `json:"error,omitempty"`

	// Name Example: postgres
	Name   string            `json:"name"`
	Status HealthCheckStatus `json:"status"`
}

// HealthCheckStatus defines model for HealthCheck.Status.
type HealthCheckStatus string

// HealthEnvelope defines model for HealthEnvelope.
type HealthEnvelope struct {
	Data  Health     `json:"data"`
	Error *ErrorBody `json:"error,omitempty"`
	Meta  *PageMeta  `json:"meta,omitempty"`

	// RequestId 请求 ID，与响应头 X-Request-ID 一致，便于排查
	RequestId string `json:"requestId"`
	Success   bool   `json:"success"`
}

// PageMeta defines model for PageMeta.
type PageMeta struct {
	Limit int   `json:"limit"`
	Page  int   `json:"page"`
	Total int64 `json:"total"`
}

// SystemInfo defines model for SystemInfo.
type SystemInfo struct {
	BuildTime string `json:"buildTime"`
	Commit    string `json:"commit"`

	// Name Example: jikelog-api
	Name       string    `json:"name"`
	ServerTime time.Time `json:"serverTime"`

	// Version Example: 0.1.0
	Version string `json:"version"`
}

// SystemInfoEnvelope defines model for SystemInfoEnvelope.
type SystemInfoEnvelope struct {
	Data  SystemInfo `json:"data"`
	Error *ErrorBody `json:"error,omitempty"`
	Meta  *PageMeta  `json:"meta,omitempty"`

	// RequestId 请求 ID，与响应头 X-Request-ID 一致，便于排查
	RequestId string `json:"requestId"`
	Success   bool   `json:"success"`
}

// Error defines model for Error.
type Error = ErrorEnvelope

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetSystemInfo 服务版本信息
	// (GET /api/v1/system/info)
	GetSystemInfo(c *gin.Context)
	// GetHealthz 存活探针
	// (GET /healthz)
	GetHealthz(c *gin.Context)
	// GetReadyz 就绪探针
	// (GET /readyz)
	GetReadyz(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetSystemInfo operation middleware
func (siw *ServerInterfaceWrapper) GetSystemInfo(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetSystemInfo(c)
}

// GetHealthz operation middleware
func (siw *ServerInterfaceWrapper) GetHealthz(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetHealthz(c)
}

// GetReadyz operation middleware
func (siw *ServerInterfaceWrapper) GetReadyz(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetReadyz(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/healthz", wrapper.GetHealthz)
	router.GET(options.BaseURL+"/readyz", wrapper.GetReadyz)
	router.GET(options.BaseURL+"/api/v1/system/info", wrapper.GetSystemInfo)
}

type ErrorJSONResponse ErrorEnvelope

type GetSystemInfoRequestObject struct {
}

type GetSystemInfoResponseObject interface {
	VisitGetSystemInfoResponse(w http.ResponseWriter) error
}

type GetSystemInfo200JSONResponse SystemInfoEnvelope

func (response GetSystemInfo200JSONResponse) VisitGetSystemInfoResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetSystemInfodefaultJSONResponse struct {
	Body       ErrorEnvelope
	StatusCode int
}

func (response GetSystemInfodefaultJSONResponse) VisitGetSystemInfoResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type GetHealthzRequestObject struct {
}

type GetHealthzResponseObject interface {
	VisitGetHealthzResponse(w http.ResponseWriter) error
}

type GetHealthz200JSONResponse HealthEnvelope

func (response GetHealthz200JSONResponse) VisitGetHealthzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetHealthzdefaultJSONResponse struct {
	Body       ErrorEnvelope
	StatusCode int
}

func (response GetHealthzdefaultJSONResponse) VisitGetHealthzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type GetReadyzRequestObject struct {
}

type GetReadyzResponseObject interface {
	VisitGetReadyzResponse(w http.ResponseWriter) error
}

type GetReadyz200JSONResponse HealthEnvelope

func (response GetReadyz200JSONResponse) VisitGetReadyzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetReadyz503JSONResponse HealthEnvelope

func (response GetReadyz503JSONResponse) VisitGetReadyzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)
	_, err := buf.WriteTo(w)
	return err
}

type GetReadyzdefaultJSONResponse struct {
	Body       ErrorEnvelope
	StatusCode int
}

func (response GetReadyzdefaultJSONResponse) VisitGetReadyzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetSystemInfo 服务版本信息
	// (GET /api/v1/system/info)
	GetSystemInfo(ctx context.Context, request GetSystemInfoRequestObject) (GetSystemInfoResponseObject, error)
	// GetHealthz 存活探针
	// (GET /healthz)
	GetHealthz(ctx context.Context, request GetHealthzRequestObject) (GetHealthzResponseObject, error)
	// GetReadyz 就绪探针
	// (GET /readyz)
	GetReadyz(ctx context.Context, request GetReadyzRequestObject) (GetReadyzResponseObject, error)
}

type StrictHandlerFunc func(ctx *gin.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictGinServerOptions struct {
	// RequestErrorHandlerFunc is called when a request cannot be parsed or
	// decoded. It is invoked for JSON bind failures, form parse/bind errors,
	// multipart reader errors, media type parse errors, missing multipart
	// boundaries, and request body read errors. The default returns 400.
	RequestErrorHandlerFunc func(ctx *gin.Context, err error)
	// HandlerErrorFunc is called when the application handler (or any
	// middleware wrapping it) returns a non-nil error. The default returns 500.
	HandlerErrorFunc func(ctx *gin.Context, err error)
	// ResponseErrorHandlerFunc is called when the response object fails to
	// serialize (Visit*Response returns an error) or when the handler returns
	// an unexpected response type. The default returns 500.
	ResponseErrorHandlerFunc func(ctx *gin.Context, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictGinServerOptions{
		RequestErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
		},
		HandlerErrorFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		},
		ResponseErrorHandlerFunc: func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictGinServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusBadRequest, gin.H{"msg": err.Error()})
		}
	}
	if options.HandlerErrorFunc == nil {
		options.HandlerErrorFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(ctx *gin.Context, err error) {
			ctx.JSON(http.StatusInternalServerError, gin.H{"msg": err.Error()})
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictGinServerOptions
}

// GetSystemInfo operation middleware
func (sh *strictHandler) GetSystemInfo(ctx *gin.Context) {
	var request GetSystemInfoRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetSystemInfo(ctx, request.(GetSystemInfoRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetSystemInfo")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetSystemInfoResponseObject); ok {
		if err := validResponse.VisitGetSystemInfoResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetHealthz operation middleware
func (sh *strictHandler) GetHealthz(ctx *gin.Context) {
	var request GetHealthzRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetHealthz(ctx, request.(GetHealthzRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetHealthz")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetHealthzResponseObject); ok {
		if err := validResponse.VisitGetHealthzResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetReadyz operation middleware
func (sh *strictHandler) GetReadyz(ctx *gin.Context) {
	var request GetReadyzRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetReadyz(ctx, request.(GetReadyzRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetReadyz")
	}

	response, err := handler(ctx, request)

	if err != nil {
		sh.options.HandlerErrorFunc(ctx, err)
	} else if validResponse, ok := response.(GetReadyzResponseObject); ok {
		if err := validResponse.VisitGetReadyzResponse(ctx.Writer); err != nil {
			sh.options.ResponseErrorHandlerFunc(ctx, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(ctx, fmt.Errorf("unexpected response type: %T", response))
	}
}